+++ src/main/java/org/example/client/GameClient.java
@@ -1,8 +1,8 @@
-package org.example;
+package org.example.client;
 
 import org.example.client.ui.screen.Window;
 
-public class Main {
+public class GameClient {
     public static void main(String[] args) {
         Window window = Window.instance;
         window.run();
