+++ src/main/java/org/example/client/communication/ServerLocation.java
@@ -1,4 +1,4 @@
-package org.example.client;
+package org.example.client.communication;
 
 import java.net.InetAddress;
 import java.util.Objects;
