+++ src/main/java/org/example/client/ui/screen/Window.java
@@ -1,6 +1,6 @@
 package org.example.client.ui.screen;
 
-import org.example.client.ServerLocation;
+import org.example.client.communication.ServerLocation;
 import org.example.client.ui.listener.KeyListener;
 import org.example.client.ui.listener.MouseListener;
 import org.example.client.ui.listener.WindowListener;
@@ -16,7 +16,7 @@ import static org.lwjgl.system.MemoryUtil.NULL;
 public class Window {
     private static final int width = 1280;
     private static final int height = 720;
-    private final String title = "Game";
+    private static final String title = "Game";
     private long glfwWindow;
     public float r, g, b, a;
 
@@ -131,8 +131,8 @@ public class Window {
         }
     }
 
-    public void setServerLocation(ServerLocation serverLocation) {
-        this.serverLocation = serverLocation;
+    public static void setServerLocation(ServerLocation serverLocation) {
+        Window.serverLocation = serverLocation;
     }
 
     public static ServerLocation getServerLocation() {
